Time: 145 ms (36.94%), Space: 13.2 MB (56.06%) - LeetHub

+++ 18-4sum/18-4sum.cpp
@@ -31,7 +31,7 @@ public:
                             right--;
                         }
                     }
-                    else if(nums[left] + nums[right] < new_target) {
+                    else if(static_cast<long long>(nums[left]) + static_cast<long long>(nums[right]) < new_target) {
                         left++;
                     }
                     else {
